I have changed the request name. As it is not good practice to keep two request same and change the case only (For ex: we have CreateLead & createLead). I think it creates ambiguity.
The best practice(as far as I learned) is to use "New" prefix for all the request if you are using the form only to create a record.
Although if you are using the same form for create & update then you should add "Edit" prefix before each request.

SFA will be bit large very soon so its safe to remove such issues in early stage.
Here is the code snippet :

    <request-map uri="CreateLead">
        <security https="true" auth="true"/>
        <response name="success" type="view" value="CreateLead"/>
    </request-map>
    <request-map uri="createLead">
        <security https="true" auth="true"/>
        <event type="service" invoke="createLead"/>
        <response name="success" type="request" value="viewprofile"/>
        <response name="error" type="view" value="CreateLead"/>
    </request-map>


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/marketing@767904 13f79535-47bb-0310-9956-ffa450edef68