A patch from Willem Janssen "Error in service signUpForContactList" (https://issues.apache.org/jira/browse/OFBIZ-4001) - OFBIZ-4001
The service signUpForContactList does not store the email address of the new entry to the list, instead it uses the email address of the first partyId NA to store to the list.

The service signUpForContactList calls service createContactListParty that has a check on a missing contactMechId. However this parameter is not a parameter that is used by the signUpForContactList service. So the check on the contactMechId has always the same result, it is missing!

signUpForContactList uses parameter prefferedContactMechId so I made a patch that adds the check on this parameter to createContactListParty. 


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/marketing@1037567 13f79535-47bb-0310-9956-ffa450edef68