A slightly modified patch from Gil Portenseigne for "Missing PartyCountryCode uiLabel" https://issues.apache.org/jira/browse/OFBIZ-5138
The label ${uiLabelMap.PartyCountryCode} is used several times in the project but i could not find the definition of PartyCountryCode in label definition files.
I looked further and found CommonCountryCode in label file which is not used and fill the need for PartyCountryCode. To solve this I propose to rename all PartyCountryCode to CommonCountryCode.

jleroux: I also moved CommonCountryCode, CommonCountryCodeMissing and CommonCountryMissing from PartyUiLabels.xml to CommonUiLabels.xml





git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/marketing@1447664 13f79535-47bb-0310-9956-ffa450edef68