Fix a bug in createLead service as described in OFBIZ-6703
Fix a bug in createLead service in which a system user cannot create
a lead more than one time in the SFA screens. The reason is that
the createLead service tries everytime to create a PartyRole of type
OWNER in this service even if it already exists which leads to a failure
and transaction rollback for the entire call.

REF https://issues.apache.org/jira/browse/OFBIZ-6703


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/marketing@1711020 13f79535-47bb-0310-9956-ffa450edef68